Moved metrics.py to pipeline
